fix output bugs
Fix the missing output of permafrost temperature both in code and
configure file.

Tested for Time-Series and GeoGrid. NOTE: GeoGrid is only tested for
single year.

add a time variable when it writes output to netcdf file.

Right now, it depends on “start_year” and “end_year” options in
configure file. In other words, the time is from “start_year” to
“end_year” by an increment of 1 yr.