Accept flattened values in set_value
This is tricky. The input to `set_value` is a flattened array;
however, the settee variable is dimensional. We don't want to clobber
the reference to the settee by changing its dimensionality.  Luckily
the `flat` iterator allows the reference to be kept, but the new
values to be fed in as a one-dimensional array with C ordering.

Thanks to @mcflugen for helping me with this.