move "Extract_Soil_Texture_Loops_New"
move the function "Extract_Soil_Texture_Loops_New" from initialize. It is because that when we intent to coupling with the Data component, we have to set spatial domination from the data obtained from data component. Then, we need to extract soil parameters again. Thus, move this function to the "update" process.